feat: Parameterize build & deploy workflow
This commit updates the `test_deploy.yml` GitHub Actions workflow to be manually triggered with `workflow_dispatch`.

It introduces `flavor` and `buildType` as inputs to allow for building and deploying different variants of the application. The Gradle build task and the Firebase App ID are now dynamically selected based on these inputs.